+++ CHANGELOG.md
@@ -1,6 +1,15 @@
 # Change Log
 
+- [1.1.0 February 28, 2018](#110-february-28-2018)
 - [1.0.0 November 6, 2017](#100-november-6-2017)
+
+## 1.1.0 February 28, 2018
+
+- Added second parameter `array $keys = []` to the following methods:
+  - `PhpCalendarI18n::getMonthNames()`
+  - `PhpCalendarI18n::getWeekDayNames()`
+  - `PhpCalendarI18n::getWeekDayAbbrs2()`
+  - `PhpCalendarI18n::getWeekDayAbbrs3()`
 
 ## 1.0.0 November 6, 2017